Projects/Chat with PDF (Streamlit; 1 file)/pyproject.toml
[project]
name = "pdf-chat-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Odessit007",email = "igor.vustyanyuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.18,<0.4.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "streamlit (>=1.42.0,<2.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain-google-genai (>=2.0.9,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
